refactor: split renderer diagnostics tests
